Big Update
Implemented parts of these classes: GameLogic, ESPSteering, PhoneSteering. All of them manage different parts of the same MQTT instance singleton. ESPSteering parses the mpu String and makes it accessible through getters. Phone Steering is empty as of now, but should make the same values accessible, so that Gamelogic only needs one set of methods, that can handle values from both Steering Classes. Second Fragment doesn't work atm, need to initiate a GameLogic instance, with ESPSteering and PhoneSteering instances inside of it. That way, the connection is made in GameLogic in the constructor, ESPSteering handles mpu data in the background and PhoneSteering does whatever. Temp and finished are handled in GameLogic