player wins in vicinity of finish
I changed the winning condition, so that the player can win, even when not moving into the direction of the finish, but rather by being close to it. This allows the players velocity to be great and the user doesn't have to try their hardest to stop at the exact right moment to move in the direction of the finish. Example: ____________Finish_______________